Moved and Updated Files
I created a new repo because I had to update the name of a parent folder, which created a number of issues. Creating a new repo was easier than trying to change the location. I also updated all of my files to fine tune them before submitting the project to Treehouse.